# Optimize rendering of ChatMessage