… panel/dist/css/chat_message.css