Attempt at improving https://github.com/holoviz/panel/issues/6064

```python
%%timeit
pn.chat.ChatMessage('Some content', user='user')
```

Before:

> 19.6 ms ± 591 µs per loop (mean ± std. dev. of 7 runs, 10 loops each

After:

> 10.9 ms ± 138 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

The main performance gains here came from avoiding use of `ParamMethod` and instead favoring manual watch callbacks.

Note those numbers also already account for the changes in: https://github.com/holoviz/param/pull/893

Without that it goes from:

> 21.6 ms ± 974 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

to:

> 11.7 ms ± 280 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)