@ahuang11 2023-12-18
There seems to be some weird indentation after these changes:
<img width="638" alt="image" src="https://github.com/holoviz/panel/assets/15331990/98a9f560-de3f-4b55-922a-0735a8f19714">

After removing width 100% - 15px, it's back to normal
<img width="1038" alt="image" src="https://github.com/holoviz/panel/assets/15331990/cdd41d29-5ff0-4d0a-96ea-08ea49129e80">