@ahuang11 2023-12-18
Some flickering seems to occur again:

https://github.com/holoviz/panel/assets/15331990/bea6b92b-15d3-4ff6-97bf-d3461250a015

On main:
https://github.com/holoviz/panel/assets/15331990/97ad3453-48a1-41f9-8607-b5caa13293f1